v0.0.1

+++ src/physai/robots/so101/ros2_node.py
@@ -9,12 +9,12 @@ from typing import Any
 import mujoco
 import numpy as np
 
-from ...bridge.messages import ROS2MessageCodec
 from ...bridge.cartesian import (
     CartesianTargetRequest,
     CartesianTargetResult,
     CartesianTargetService,
 )
+from ...bridge.messages import ROS2MessageCodec
 from ...bridge.mujoco_ros_bridge import MuJoCoROSBridge, RclpyTransport
 from .env import EnvConfig, SO101Env
 
@@ -249,7 +249,7 @@ class SO101ROS2Node:
 def _copy_ros_stamp(header: Any, stamp: float) -> None:
     seconds = max(0.0, float(stamp))
     header.stamp.sec = int(seconds)
-    header.stamp.nanosec = int(round((seconds - header.stamp.sec) * 1e9))
+    header.stamp.nanosec = round((seconds - header.stamp.sec) * 1e9)
 
 
 __all__ = ["SO101ROS2Node"]
